django_backend/requirements.txt
Django
djangorestframework
clickhouse-connect
django-cors-headers